refactor: harden and modularize overlay monitor
