+++ package.json
@@ -14,6 +14,8 @@
     "eft:story": "mkdir -p data/eft && tsx scripts/eft-story-wiki.ts && tsx scripts/eft-story-generate.ts > data/eft/story-final.json && tsx scripts/eft-story-write.ts",
     "wiki:compare": "tsx scripts/wiki-compare.ts",
     "monitor": "node monitor/server.js",
+    "check:monitor": "node --check monitor/server.js && node --check monitor/lib/config.js && node --check monitor/lib/sections.js && node --check monitor/public/app.js",
+    "audit:dependencies": "npm audit --omit=optional --audit-level=high",
     "typecheck": "tsc --noEmit",
     "format": "prettier --write \"{scripts,src/lib,tests,examples}/**/*.ts\"",
     "format:check": "prettier --check \"{scripts,src/lib,tests,examples}/**/*.ts\"",
